+++ deno_dist/jsx/dom/jsx-dev-runtime.ts
@@ -1,4 +1,4 @@
-import type { Props } from '../index.ts'
+import type { Props } from '../base.ts'
 import { normalizeIntrinsicElementProps } from '../utils.ts'
 
 export const jsxDEV = (tag: string | Function, props: Props, key: string | undefined) => {
